Add iffy Haskell implementation of part 2 day 2
Its way too late for me to be doing this... I should have wrote this in
a new file as I did with day one, but I'm too tired to do that. I'm
committing this now, but I plan to roll it back and do the right thing
tomorrow. :shrug: